better use of workspaces; kind of fixed viper-battery custom module
Seems like I couldn't just run razer-cli on startup, so I just instead
call python to run the script and fetch the battery level. It starts at 0
but updates on next interval, so I'm okay with that for now.
Workspaces now have the first one bound to the main monitor (as should have
been always) and workspaces 2 and 3 are bound to the second monitor.
Also cleaned up files and dead code in the config.