Tests and DAP working (i think), so done for today
Major problem was that the vscode-test runner that is in the mason registry is outdated
so it could not run any tests.Had to build an updated version from source, but I think
it is currently working nicely now.
I think this works OK, but maybe mason-nvim-dap is overriding what I am trying to
set up through nvim-jdtls. Will have to check that sometime.
Anyways, pretty happy with how this is set up now. I hope I can forget about the config
for a long time soon.